+++ tests/cli_commands/test_run_test.py
@@ -16,7 +16,7 @@ class RunTestCommandTestCase(unittest.TestCase):
     def setUp(self) -> None:
         self.subprocess_run_args = []
 
-    def _mock_run(self, args: List[str]):
+    def _mock_run(self, args: List[str], **_kwargs):
         self.subprocess_run_args = args
 
     @patch(